AI_Powered_Chat_Bot_System/backend/image_processor.py
import cv2
import tensorflow as tf
from tensorflow.keras.applications import VGG16

class ImageProcessor:
    def __init__(self):
        self.model = VGG16(weights='imagenet')

    def describe_image(self, image_content):
        # Code to process image content and generate description
        image = self.preprocess_image(image_content)
        description = self.model.predict(image)
        return "Description of the image"

    def preprocess_image(self, image_content):
        # Implement image preprocessing logic here
        return image_content
