AI_Powered_Chat_Bot_System/requirements.txt
fastapi
uvicorn
pymongo
pdfminer.six
opencv-python
tensorflow
beautifulsoup4
requests
openai
llama-index
python-dotenv
